Allow default media controls to be hidden.
