+++ src/localize/languages/pt-BR.json
@@ -84,6 +84,7 @@
     },
     "common": {
       "controls": {
+        "builtin": "",
         "filter": {
           "editor_label": "Filtro de Mídia",
           "mode": "Modo do filtro",
