+++ src/patches/ha-web-rtc-player.ts
@@ -72,9 +72,9 @@ customElements.whenDefined('ha-web-rtc-player').then(() => {
       }
     }
 
-    public async setControls(controls: boolean): Promise<void> {
+    public async setControls(controls?: boolean): Promise<void> {
       if (this._video) {
-        this._video.controls = controls;
+        this._video.controls = controls ?? this.controls;
       }
     }
 
@@ -100,7 +100,12 @@ customElements.whenDefined('ha-web-rtc-player').then(() => {
           ?playsinline=${this.playsInline}
           ?controls=${this.controls}
           @loadedmetadata=${() => {
-            hideMediaControlsTemporarily(this._video, MEDIA_LOAD_CONTROLS_HIDE_SECONDS);
+            if (this.controls) {
+              hideMediaControlsTemporarily(
+                this._video,
+                MEDIA_LOAD_CONTROLS_HIDE_SECONDS,
+              );
+            }
           }}
           @loadeddata=${(ev) => {
             dispatchMediaLoadedEvent(this, ev, {
